Implement support for async and iterators. (#3)
* Implement support for async and iterators.

This change makes the decorator work correctly for async functions,
instead of always returning the same coroutine, which can only be
awaited once and calls subsequent calls to fail.

This also detects returned iterators, evaluates them to completion, and
returns the result as a tuple. Prior to this change, an exhausted
iterator would be returned.

* Explicitly disable async iterators.

Also, better handle sync iterators to be API-compatible with the
unwrapped function call by returning an iterator object instead of a
tuple object.